Update content/examples/Topics/Fractals and L-Systems/Mandelbrot/Mandelbrot.pde

+++ content/examples/Topics/Fractals and L-Systems/Mandelbrot/Mandelbrot.pde
@@ -48,7 +48,7 @@ for (int j = 0; j < height; j++) {
     float a = x;
     float b = y;
     int n = 0;
-    double max = 4.0;  // Infinty in our finite world is simple, let's just consider it 4
+    float max = 4.0;  // Infinity in our finite world is simple, let's just consider it 4
     double absOld = 0.0;
     float convergeNumber = maxiterations; // this will change if the while loop breaks due to non-convergence
     while (n < maxiterations) {
